Add invalid content error (#24)
